# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TECA API",
    "description": "\nAPI del e-commerce de muebles **TECA — Para toda la vida**.\n\nCubre la tienda pública (catálogo, carrito, checkout de invitado o con cuenta),\nel área del cliente (pedidos, direcciones, métodos de pago, devoluciones) y el\npanel de administración (usuarios internos, productos, pedidos, cupones,\nauditoría y finanzas).\n\n### Autenticación\nTodas las rutas protegidas usan **JWT Bearer**:\n\n1. `POST /auth/login` con correo y contraseña.\n2. Copia el `access_token` de la respuesta.\n3. Pulsa **Authorize** (arriba a la derecha) y pégalo.\n\n### Roles\n| Rol | Alcance |\n|---|---|\n| `cliente` | Tienda: carrito, pedidos propios, cuenta, reseñas |\n| `admin` | Acceso total al panel |\n| `editor` | Productos |\n| `encargado` | Productos, pedidos y devoluciones |\n| `vendedor` | Pedidos (y lectura de productos) |\n| `soporte` | Devoluciones y lectura de pedidos |\n| `finanzas` | Reportes financieros |\n\n### Convenciones\n- Moneda: balboas (B/.), equivalente a USD.\n- Los identificadores son `ObjectId` de Mongo expuestos como `id` (string).\n- Errores en formato `{\"detail\": \"mensaje\"}` con el código HTTP correspondiente.\n",
    "version": "1.0.0"
  },
  "paths": {
...
    "/cart": {
      "get": {
        "tags": [
          "Carrito"
        ],
        "summary": "Ver mi carrito",
        "description": "Pantalla **Carrito de compra**. Devuelve los items con datos frescos del producto (nombre, precio, material, imagen, stock y subtotal de línea) más el bloque *Resumen del pedido*:\n\n| Campo | Uso en la pantalla |\n|---|---|\n| `subtotal` | Suma de las líneas |\n| `shipping` | Envío (gratis al superar el umbral configurado) |\n| `missing_for_free_shipping` | *«Te faltan B/. XX.XX para envío gratis»* |\n| `discount` / `coupon_code` | Línea de descuento del cupón aplicado |\n| `total` | Total a pagar |\n\nSi el carrito no existe todavía, se crea vacío.\n\n🔒 Requiere token.",
        "operationId": "get_cart_cart_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/cart/items": {
      "post": {
        "tags": [
          "Carrito"
        ],
        "summary": "Agregar un producto al carrito",
        "description": "Botón *Agregar al carrito*. Si el producto ya estaba en el carrito, **suma** la cantidad en lugar de duplicar la línea.\n\nValida el stock contra la cantidad resultante y devuelve el carrito completo ya recalculado, para que el frontend refresque el resumen con una sola llamada.\n\n🔒 Requiere token. **Errores:** `404` producto inexistente · `409` si se supera el stock disponible.",
        "operationId": "add_item_cart_items_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Item del carrito"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Producto no encontrado"
          },
          "409": {
            "description": "No puedes seleccionar más de N unidades"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error de validación"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/cart/items/{product_id}": {
      "put": {
        "tags": [
          "Carrito"
        ],
        "summary": "Cambiar la cantidad de un producto",
        "description": "Fija la cantidad exacta de una línea del carrito (selector `- 1 +`). A diferencia de `POST /cart/items`, **reemplaza** la cantidad en vez de sumarla.\n\nDevuelve el carrito recalculado.\n\n🔒 Requiere token. **Errores:** `404` si el producto no existe o no está en el carrito · `409` si la cantidad supera el stock.",
        "operationId": "update_item_cart_items__product_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "ID del producto que ya está en el carrito",
              "title": "Product Id"
            },
            "description": "ID del producto que ya está en el carrito"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Item del carrito"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "El producto no está en el carrito"
          },
          "409": {
            "description": "No puedes seleccionar más de N unidades"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error de validación"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Carrito"
        ],
        "summary": "Quitar un producto del carrito",
        "description": "Icono de papelera de la fila. Elimina la línea completa sin importar la cantidad y devuelve el carrito recalculado. Es idempotente: si el producto ya no estaba, responde igual sin error.\n\n🔒 Requiere token.",
        "operationId": "remove_item_cart_items__product_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "ID del producto a quitar del carrito",
              "title": "Product Id"
            },
            "description": "ID del producto a quitar del carrito"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error de validación"
                }
              }
            }
          }
        }
      }
    },
    "/cart/coupon": {
      "post": {
        "tags": [
          "Carrito"
        ],
        "summary": "Aplicar un cupón de descuento",
        "description": "Aplica un código de descuento al carrito (ej. `TECA10` = 10 %). El código no distingue mayúsculas.\n\n**Solo se admite un cupón a la vez:** si ya había uno aplicado, este lo reemplaza (el aviso *«Si aplicas otro, el cupón actual será reemplazado»*).\n\nDevuelve el carrito con la línea `discount` y el `total` actualizados.\n\n🔒 Requiere token. **Errores:** `404` si el cupón no existe o está inactivo.",
        "operationId": "apply_coupon_cart_coupon_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Cupón a aplicar"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Cupón inválido o expirado"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error de validación"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Carrito"
        ],
        "summary": "Quitar el cupón aplicado",
        "description": "Botón *Eliminar* del recuadro «Código de cupón aplicado». Deja el descuento en 0 y devuelve el carrito recalculado.\n\n🔒 Requiere token.",
        "operationId": "remove_coupon_cart_coupon_delete",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Cupón": {
        "properties": {
          "code": {
            "type": "string",
            "maxLength": 20,
            "minLength": 3,
            "title": "Código",
            "description": "Se guarda en mayúsculas y debe ser único"
          },
          "discount_percent": {
            "type": "number",
            "maximum": 100.0,
            "exclusiveMinimum": 0.0,
            "title": "Descuento (%)",
            "description": "Porcentaje sobre el subtotal"
          },
          "active": {
            "type": "boolean",
            "title": "Activo",
            "description": "Ponlo en falso para desactivarlo sin borrarlo",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "code",
          "discount_percent"
        ],
        "title": "Cupón",
        "example": {
          "active": true,
          "code": "TECA10",
          "discount_percent": 10
        }
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}